pyproject.toml
[project]
name = "dnd"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.28.1",
    "google-cloud-aiplatform>=1.145.0",
    "chromadb>=1.0.0",
    "pypdf>=6.9.2",
    "python-dotenv>=1.2.2",
]
